# Conflicts:
#	DungeonCrawler/DungeonCrawler/Debug/DungeonCrawler.tlog/CL.read.1.tlog
#	DungeonCrawler/DungeonCrawler/Debug/DungeonCrawler.tlog/CL.write.1.tlog
#	DungeonCrawler/DungeonCrawler/Debug/DungeonCrawler.tlog/cl.command.1.tlog
#	DungeonCrawler/DungeonCrawler/Debug/DungeonCrawler.tlog/link.command.1.tlog
#	DungeonCrawler/DungeonCrawler/Debug/DungeonCrawler.tlog/link.read.1.tlog
#	DungeonCrawler/DungeonCrawler/Debug/DungeonCrawler.tlog/rc.command.1.tlog
#	DungeonCrawler/DungeonCrawler/Debug/vc140.idb
#	DungeonCrawler/DungeonCrawler/Level.cpp
#	DungeonCrawler/DungeonCrawler/Level.h

+++ DungeonCrawler/DungeonCrawler/Game.h
@@ -10,7 +10,7 @@ class Game
 private:
 	Level* currentLevel;
 	int levelID = 0;
-	int currentLevelID;
+	size_t currentLevelID;
 	Player p1 = Player("Dave");
 	std::vector<Level*> *levels;
 public:
